+++ src/main/java/kr/co/moneybridge/model/reservation/ReviewRepository.java
@@ -14,7 +14,7 @@ import java.util.Optional;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     @Query(value = "select count(r) from Review r where r.reservation.pb.id = :pbId")
-    Integer countByPBId(@Param("pbId") Long pbId);
+    Long countByPBId(@Param("pbId") Long pbId);
 
     @Modifying
     @Query("delete from Review r where r.reservation.id = :reservationId")
